deleting nodemodules folder
to use the application, this folder is required but it increases the size while sharing so i have uninstalled it.
to use it just open the folder in your drive
and while you have cd to the folder in terminal enter 'npm install' in your terminal and the node modules folder will be installed.